task: complete cat javascript implementation
