+++ implement-shell-tools/package.json
@@ -1,0 +1,5 @@
+{
+	"dependencies": {
+		"commander": "^15.0.0"
+	}
+}
